+++ inc/ti/fn/fntest.h
@@ -1,13 +1,18 @@
 #include <ti/fn/fn.h>
 
-static int do__f_test(ti_query_t * query, cleri_node_t * nd, ex_t * e)
+static int test__deprecated_on_string(
+        ti_query_t * query,
+        cleri_node_t * nd,
+        ex_t * e)
 {
     const int nargs = langdef_nd_n_function_params(nd);
     ti_raw_t * raw;
     _Bool has_match;
 
-    if (!ti_val_is_str(query->rval))
-        return fn_call_try("test", query, nd, e);
+    log_warning(
+            "function `test()` on type `str` is deprected; "
+            "use the function on type `regex` instead "
+            "with the string to test as argument");
 
     if (fn_nargs("test", DOC_STR_TEST, 1, nargs, e))
         return e->nr;
@@ -33,5 +38,44 @@ static int do__f_test(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
 failed:
     ti_val_unsafe_drop((ti_val_t *) raw);
+    return e->nr;
+
+}
+
+static int do__f_test(ti_query_t * query, cleri_node_t * nd, ex_t * e)
+{
+    const int nargs = langdef_nd_n_function_params(nd);
+    ti_regex_t * regex;
+    _Bool has_match;
+
+    if (!ti_val_is_regex(query->rval))
+        return ti_val_is_str(query->rval)
+                ? test__deprecated_on_string(query, nd, e)
+                : fn_call_try("test", query, nd, e);
+
+    if (fn_nargs("test", DOC_REGEX_TEST, 1, nargs, e))
+        return e->nr;
+
+    regex = (ti_regex_t *) query->rval;
+    query->rval = NULL;
+
+    if (ti_do_statement(query, nd->children->node, e))
+        goto failed;
+
+    if (!ti_val_is_str(query->rval))
+    {
+        ex_set(e, EX_TYPE_ERROR,
+            "function `test` expects argument 1 to be "
+            "of type `"TI_VAL_STR_S"` but got type `%s` instead"DOC_STR_TEST,
+            ti_val_str(query->rval));
+        goto failed;
+    }
+
+    has_match = ti_regex_test(regex, (ti_raw_t *) query->rval);
+    ti_val_unsafe_drop(query->rval);
+    query->rval = (ti_val_t *) ti_vbool_get(has_match);
+
+failed:
+    ti_val_unsafe_drop((ti_val_t *) regex);
     return e->nr;
 }